1. Remove the unnecessary typescript files 2. Change the graphql code generate config

+++ backend/src/main/java/com/opencrm/app/model/BaseEntity.java
@@ -1,10 +1,24 @@
 package com.opencrm.app.model;
 
+import java.io.Serializable;
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+
 @MappedSuperclass
 public class BaseEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @CreatedDate
     @Column(updatable = false)
@@ -22,11 +36,11 @@ public class BaseEntity implements Serializable {
     @Column(insertable = false)
     private Integer updatedBy;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
